# src/main/resources/static/swagger-config.yml
openapi: 3.0.1
info:
  title: Gym Service
  version: v0.1
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /api/v1/trainee:
    get:
      tags:
        - Trainee service
      summary: Get trainee by username
      description: This can only be executed after authentication
      operationId: getTraineeByUsername
      parameters:
        - name: username
          in: query
          description: The username that needs to be fetched
          required: true
          example: Nick.Trainee
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraineeWithTrainersDTO'
        '400':
          description: Specified wrong fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trainee with this username doesnt found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Trainee service
      summary: Update trainee
      description: This can only be executed after authentication
      operationId: updateTrainee
      requestBody:
        description: Details to update trainee
        content:
          application/json:
            schema:
              type: object
              properties:
                userName:
                  description: Trainee username (required)
                  type: string
                  example: Nick.Trainee
                firstName:
                  description: Trainee firstName (required)
                  type: string
                  example: Nick
                lastName:
                  description: Trainee lastName (required)
                  type: string
                  example: Trainee
                dateOfBirth:
                  type: string
                  format: date
                  description: Trainee date of birth (optional)
                  example: 2024-01-01
                address:
                  type: string
                  description: Trainee address (optional)
                isActive:
                  description: Trainee isActive status (required)
                  type: boolean
                  example: true
              required:
                - userName
                - firstName
                - lastName
                - isActive
        required: true
      responses:
        '200':
          description: Trainee updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraineeWithTrainersDTO'
        '400':
          description: Specified wrong fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trainee with this username doesnt found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Trainee service
      summary: Create a new trainee
      operationId: createTrainee
      requestBody:
        description: Trainee details
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  description: Trainee firstName (required)
                  type: string
                  example: Nick
                lastName:
                  description: Trainee lastName (required)
                  type: string
                  example: Trainee
                dateOfBirth:
                  description: Trainee dateOfBirth (optional)
                  type: string
                  format: date
                  example: 2024-03-12
                address:
                  description: Trainee address (optional)
                  type: string
                  example: Maple street
              required:
                - firstName
                - lastName
        required: true
      responses:
        '201':
          description: Trainee created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCredentialsDTO'
        '400':
          description: Specified wrong fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Trainee service
      summary: Delete trainee
      description: This can only be executed after authentication
      operationId: deleteTrainee
      parameters:
        - name: username
          in: query
          description: The username that needs to be fetched
          required: true
          example: Nick.Trainee
          schema:
            type: string
      responses:
        '200':
          description: Trainee deleted successfully
        '400':
          description: Specified wrong fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trainee with this username doesnt found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags:
        - Trainee service
      summary: Change trainee status
      description: This can only be executed after authentication
      operationId: changeTraineeStatus
      parameters:
        - name: username
          in: query
          description: Trainee username
          required: true
          example: Nick.Trainee
          schema:
            type: string
        - name: isActive
          in: query
          description: Change isActive status for Trainee
          required: true
          example: true
          schema:
            type: boolean
      responses:
        '200':
          description: Trainee status changed successfully
        '400':
          description: Specified wrong fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trainee with this username doesnt found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/trainee/trainers:
    put:
      tags:
        - Trainee service
      summary: Update trainee's trainers list
      description: This can only be executed after authentication
      operationId: updateTraineesTrainersList
      requestBody:
        description: Trainee with trainers details
        content:
          application/json:
            schema:
              required:
                - trainers
                - userName
              type: object
              properties:
                userName:
                  type: string
                  description: Trainee userName (required)
                trainers:
                  type: array
                  items:
                    required:
                      - userName
                    type: object
                    properties:
                      userName:
                        type: string
                        description: Trainer userName (required)
                    description: Trainee List of Trainers userNames that must be updated (required)
        required: true
      responses:
        '200':
          description: Trainee's trainers list updated successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainerDTO'
        '400':
          description: Specified wrong fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trainee with this username doesnt found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/trainee/trainings:
    get:
      tags:
        - Trainee service
      summary: Get trainee trainings by criteria
      description: This can only be executed after authentication
      operationId: getTraineeTrainingsByCriteria
      parameters:
        - name: username
          in: query
          required: true
          example: Nick.Trainee
          description: The trainee username to be provided
          schema:
            type: string
        - name: fromDate
          in: query
          required: false
          example: 2024-01-01
          description: The filtering criteria by date, starting from this specific date
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          required: false
          example: 2024-02-01
          description: The filtering criteria by date, ending with this specific date
          schema:
            type: string
            format: date
        - name: trainerUsername
          in: query
          required: false
          example: John.Trainer
          description: The filtering criteria by trainee username
          schema:
            type: string
        - name: trainingTypeName
          in: query
          required: false
          example: Boxing
          description: The filtering criteria by training type
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainingDTO'
        '400':
          description: Specified wrong fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trainee with this username doesnt found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/trainee/available-trainers:
    get:
      tags:
        - Trainee service
      summary: Get trainers not assigned to trainee
      description: This can only be executed after authentication
      operationId: getTrainersNotAssignedToTrainee
      parameters:
        - name: username
          in: query
          description: The username that needs to be fetched
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainerDTO'
        '400':
          description: Specified wrong fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trainee with this username doesnt found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    TrainerDTO:
      required:
        - firstName
        - lastName
        - specialization
      type: object
      properties:
        userName:
          type: string
          description: User userName (optional)
        firstName:
          type: string
          description: User firstName (required)
        lastName:
          type: string
          description: User lastName (required)
        isActive:
          type: boolean
          description: User isActive (optional)
        specialization:
          $ref: '#/components/schemas/TrainingTypeDTO'
    TrainingTypeDTO:
      required:
        - trainingTypeName
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        trainingTypeName:
          type: string
          description: trainingTypeName (required)
...
    TraineeWithTrainersDTO:
      required:
        - firstName
        - lastName
      type: object
      properties:
        userName:
          type: string
          description: User userName (optional)
        firstName:
          type: string
          description: User firstName (required)
        lastName:
          type: string
          description: User lastName (required)
        isActive:
          type: boolean
          description: User isActive (optional)
        address:
          type: string
        dateOfBirth:
          type: string
          format: date
        trainers:
          type: array
          items:
            $ref: '#/components/schemas/TrainerDTO'
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
          example: message
        errorType:
          type: string
          description: Error type
          example: Type based on error
          enum:
            - VALIDATION_ERROR
            - PROCESSING_ERROR
            - DATABASE_ERROR
            - FATAL_ERROR
            - AUTHORIZATION_ERROR
        timeStamp:
          type: string
          format: date-time
...
    TrainingDTO:
      required:
        - duration
        - trainingDate
        - trainingName
        - trainingTypeName
      type: object
      properties:
        traineeUserName:
          type: string
        trainerUserName:
          type: string
        trainingName:
          type: string
        trainingDate:
          type: string
          format: date
        duration:
          type: integer
          format: int32
        trainingTypeName:
          type: string
    UserCredentialsDTO:
      required:
        - password
        - username
      type: object
      properties:
        username:
          type: string
        password:
          type: string